backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.2.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic